+++ common/src/main/java/dev/aika/smsn/SMSN.java
@@ -1,6 +1,8 @@
 package dev.aika.smsn;
 
 import dev.aika.smsn.config.SMSNConfig;
+import dev.aika.smsn.mixin.MixinPlatform;
+import dev.aika.smsn.mixin.ModMixinManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -10,7 +12,13 @@ public final class SMSN {
     public static final Logger LOGGER = LoggerFactory.getLogger(MOD_NAME);
 
     public static SMSNConfig CONFIG = new SMSNConfig();
+    public static final ModMixinManager MixinManager = new ModMixinManager();
 
     public static void init() {
+    }
+
+    public static void preInit() {
+        CONFIG = SMSNConfig.load();
+        MixinPlatform.register();
     }
 }