+++ common/src/main/java/dev/aika/smsn/SMSNPlatform.java
@@ -6,16 +6,6 @@ import java.nio.file.Path;
 
 public class SMSNPlatform {
     @ExpectPlatform
-    public static boolean isModLoaded(String modId) {
-        throw new AssertionError();
-    }
-
-    @ExpectPlatform
-    public static String getModName(String modId) {
-        throw new AssertionError();
-    }
-
-    @ExpectPlatform
     public static Path getConfigDir() {
         throw new AssertionError();
     }