+++ common/src/main/java/dev/aika/smsn/mixin/ModMixinInfo.java
@@ -1,7 +1,7 @@
 package dev.aika.smsn.mixin;
 
 import dev.aika.smsn.SMSN;
-import dev.aika.smsn.SMSNPlatform;
+import dev.aika.smsn.api.ModPlatform;
 import lombok.Getter;
 import org.intellij.lang.annotations.Language;
 import org.jetbrains.annotations.NotNull;
@@ -10,27 +10,27 @@ import java.util.List;
 
 public class ModMixinInfo implements Comparable<ModMixinInfo> {
     @Getter
-    private final String modid;
+    private final String modId;
     @Getter
     private final List<String> mixinClasses;
     @Getter
     private final boolean isModLoaded;
 
-    private ModMixinInfo(String modid, List<String> mixinClasses) {
-        this.modid = modid;
+    public ModMixinInfo(String modId, List<String> mixinClasses) {
+        this.modId = modId;
         this.mixinClasses = mixinClasses;
-        this.isModLoaded = SMSNPlatform.isModLoaded(modid);
+        this.isModLoaded = ModPlatform.isModLoaded(modId);
     }
 
-    public static ModMixinInfo create(String modid, @Language(value = "JAVA", suffix = ".class") String... mixinClasses) {
-        return new ModMixinInfo(modid, List.of(mixinClasses));
+    public ModMixinInfo(String modId, @Language(value = "JAVA", suffix = ".class") String... mixinClasses) {
+        this(modId, List.of(mixinClasses));
     }
 
     @Override
     public int compareTo(@NotNull ModMixinInfo o) {
-        if (modid.hashCode() < o.modid.hashCode())
+        if (modId.hashCode() < o.modId.hashCode())
             return -1;
-        else if (modid.hashCode() == o.modid.hashCode())
+        else if (modId.hashCode() == o.modId.hashCode())
             return 0;
         return 1;
     }