+++ fabric/src/main/java/dev/aika/smsn/mixin/fabric/MixinPlatformImpl.java
@@ -1,53 +1,56 @@
 package dev.aika.smsn.mixin.fabric;
 
+import dev.aika.smsn.SMSN;
 import dev.aika.smsn.mixin.ModMixinInfo;
-import dev.aika.smsn.mixin.ModMixinManager;
 
 @SuppressWarnings({"unused", "SpellCheckingInspection"})
 public class MixinPlatformImpl {
     public static void register() {
-        ModMixinManager.addInfo(
-                ModMixinInfo.create("ad_astra",
+        SMSN.MixinManager.addInfo(
+                new ModMixinInfo("ad_astra",
                         "dev.aika.smsn.fabric.mixin.ad_astra.StationLoaderMixin"
                 ),
-                ModMixinInfo.create("nitrogen",
+                new ModMixinInfo("nitrogen",
                         "dev.aika.smsn.fabric.mixin.aetherteam.UserData$ServerMixin"
                 ),
-                ModMixinInfo.create("botania",
+                new ModMixinInfo("botania",
                         "dev.aika.smsn.fabric.mixin.botania.ContributorListMixin"
                 ),
-                ModMixinInfo.create("tram_additions",
+                new ModMixinInfo("tram_additions",
                         "dev.aika.smsn.fabric.mixin.create_tram_additions.WorldTickMixin",
                         "dev.aika.smsn.fabric.mixin.create_tram_additions.TramAdditionsMixin"
                 ),
-                ModMixinInfo.create("exposure",
+                new ModMixinInfo("exposure",
                         "dev.aika.smsn.fabric.mixin.exposure.GildedMixin",
                         "dev.aika.smsn.fabric.mixin.exposure.PatreonMixin"
                 ),
-                ModMixinInfo.create("mru",
+                new ModMixinInfo("mru",
                         "dev.aika.smsn.fabric.mixin.imb11.mru.APIMixin"
                 ),
-                ModMixinInfo.create("inventoryprofilesnext",
+                new ModMixinInfo("immersive_portals_core",
+                        "dev.aika.smsn.fabric.mixin.immersiveportals.IPModInfoCheckingMixin"
+                ),
+                new ModMixinInfo("inventoryprofilesnext",
                         "dev.aika.smsn.fabric.mixin.ipn.IPNInfoManager$doCheckVersion$$inlined$timer$default$1Mixin",
                         "dev.aika.smsn.fabric.mixin.ipn.IPNInfoManager$doSessionKeepAlive$$inlined$timer$default$1Mixin"
                 ),
-                ModMixinInfo.create("iris",
+                new ModMixinInfo("iris",
                         "dev.aika.smsn.fabric.mixin.iris.UpdateCheckerMixin"
                 ),
-                ModMixinInfo.create("ribbits",
+                new ModMixinInfo("ribbits",
                         "dev.aika.smsn.fabric.mixin.ribbits.SupportersJSONMixin"
                 ),
-                ModMixinInfo.create("supplementaries",
+                new ModMixinInfo("supplementaries",
                         "dev.aika.smsn.fabric.mixin.supplementaries.CreditsMixin"
                 ),
-                ModMixinInfo.create("tenshilib",
+                new ModMixinInfo("tenshilib",
                         "dev.aika.smsn.fabric.mixin.tenshilib.PatreonDataManagerMixin"
                 ),
-                ModMixinInfo.create("xaerominimap",
+                new ModMixinInfo("xaerominimap",
                         "dev.aika.smsn.fabric.mixin.xaero.minimap.InternetMixin",
                         "dev.aika.smsn.fabric.mixin.xaero.minimap.PatreonMixin"
                 ),
-                ModMixinInfo.create("xaeroworldmap",
+                new ModMixinInfo("xaeroworldmap",
                         "dev.aika.smsn.fabric.mixin.xaero.world_map.InternetMixin",
                         "dev.aika.smsn.fabric.mixin.xaero.world_map.PatreonMixin"
                 )