+++ forge/src/main/java/dev/aika/smsn/forge/SMSNPlatformImpl.java
@@ -1,21 +1,11 @@
 package dev.aika.smsn.forge;
 
-import net.minecraftforge.fml.loading.FMLLoader;
 import net.minecraftforge.fml.loading.FMLPaths;
 
 import java.nio.file.Path;
 
 @SuppressWarnings("unused")
 public class SMSNPlatformImpl {
-    public static boolean isModLoaded(String modId) {
-        return FMLLoader.getLoadingModList().getModFileById(modId) != null;
-    }
-
-    public static String getModName(String modId) {
-        if (!isModLoaded(modId)) return null;
-        return FMLLoader.getLoadingModList().getModFileById(modId).moduleName();
-    }
-
     public static Path getConfigDir() {
         return FMLPaths.CONFIGDIR.get();
     }